BandsJitter is functional but the output is not tested

+++ dataset/EuroSAT/reader.py
@@ -23,7 +23,7 @@ class EuroSAT(Dataset):
         p = pathlib.Path(self.image_paths[idx])
         label, path = p.parent.name, p
         im = utils.read(self.dataset_root / path)
-        im = torch.from_numpy(im.astype("float16").to_numpy())
+        im = torch.from_numpy(im.astype("float32").to_numpy())
         if self.transforms:
             im = self.transforms(im)
         else:
